+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/resources/_resource_management_client.py
@@ -77,6 +77,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :mod:`v2017_05_10.models<azure.mgmt.resource.resources.v2017_05_10.models>`
            * 2018-02-01: :mod:`v2018_02_01.models<azure.mgmt.resource.resources.v2018_02_01.models>`
            * 2018-05-01: :mod:`v2018_05_01.models<azure.mgmt.resource.resources.v2018_05_01.models>`
+           * 2019-03-01: :mod:`v2019_03_01.models<azure.mgmt.resource.resources.v2019_03_01.models>`
            * 2019-05-01: :mod:`v2019_05_01.models<azure.mgmt.resource.resources.v2019_05_01.models>`
            * 2019-05-10: :mod:`v2019_05_10.models<azure.mgmt.resource.resources.v2019_05_10.models>`
            * 2019-07-01: :mod:`v2019_07_01.models<azure.mgmt.resource.resources.v2019_07_01.models>`
@@ -98,6 +99,9 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             return models
         elif api_version == '2018-05-01':
             from .v2018_05_01 import models
+            return models
+        elif api_version == '2019-03-01':
+            from .v2019_03_01 import models
             return models
         elif api_version == '2019-05-01':
             from .v2019_05_01 import models
@@ -128,6 +132,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2017_05_10.operations.DeploymentOperations>`
            * 2018-02-01: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2018_02_01.operations.DeploymentOperations>`
            * 2018-05-01: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2018_05_01.operations.DeploymentOperations>`
+           * 2019-03-01: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2019_03_01.operations.DeploymentOperations>`
            * 2019-05-01: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2019_05_01.operations.DeploymentOperations>`
            * 2019-05-10: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2019_05_10.operations.DeploymentOperations>`
            * 2019-07-01: :class:`DeploymentOperations<azure.mgmt.resource.resources.v2019_07_01.operations.DeploymentOperations>`
@@ -146,6 +151,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import DeploymentOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import DeploymentOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import DeploymentOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import DeploymentOperations as OperationClass
         elif api_version == '2019-05-10':
@@ -171,6 +178,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2017_05_10.operations.DeploymentsOperations>`
            * 2018-02-01: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2018_02_01.operations.DeploymentsOperations>`
            * 2018-05-01: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2018_05_01.operations.DeploymentsOperations>`
+           * 2019-03-01: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2019_03_01.operations.DeploymentsOperations>`
            * 2019-05-01: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2019_05_01.operations.DeploymentsOperations>`
            * 2019-05-10: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2019_05_10.operations.DeploymentsOperations>`
            * 2019-07-01: :class:`DeploymentsOperations<azure.mgmt.resource.resources.v2019_07_01.operations.DeploymentsOperations>`
@@ -189,6 +197,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import DeploymentsOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import DeploymentsOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import DeploymentsOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import DeploymentsOperations as OperationClass
         elif api_version == '2019-05-10':
@@ -210,6 +220,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
         """Instance depends on the API version:
 
            * 2018-05-01: :class:`Operations<azure.mgmt.resource.resources.v2018_05_01.operations.Operations>`
+           * 2019-03-01: :class:`Operations<azure.mgmt.resource.resources.v2019_03_01.operations.Operations>`
            * 2019-05-01: :class:`Operations<azure.mgmt.resource.resources.v2019_05_01.operations.Operations>`
            * 2019-05-10: :class:`Operations<azure.mgmt.resource.resources.v2019_05_10.operations.Operations>`
            * 2019-07-01: :class:`Operations<azure.mgmt.resource.resources.v2019_07_01.operations.Operations>`
@@ -220,6 +231,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
         api_version = self._get_api_version('operations')
         if api_version == '2018-05-01':
             from .v2018_05_01.operations import Operations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import Operations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import Operations as OperationClass
         elif api_version == '2019-05-10':
@@ -245,6 +258,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2017_05_10.operations.ProvidersOperations>`
            * 2018-02-01: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2018_02_01.operations.ProvidersOperations>`
            * 2018-05-01: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2018_05_01.operations.ProvidersOperations>`
+           * 2019-03-01: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2019_03_01.operations.ProvidersOperations>`
            * 2019-05-01: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2019_05_01.operations.ProvidersOperations>`
            * 2019-05-10: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2019_05_10.operations.ProvidersOperations>`
            * 2019-07-01: :class:`ProvidersOperations<azure.mgmt.resource.resources.v2019_07_01.operations.ProvidersOperations>`
@@ -263,6 +277,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import ProvidersOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import ProvidersOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import ProvidersOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import ProvidersOperations as OperationClass
         elif api_version == '2019-05-10':
@@ -288,6 +304,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2017_05_10.operations.ResourceGroupsOperations>`
            * 2018-02-01: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2018_02_01.operations.ResourceGroupsOperations>`
            * 2018-05-01: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2018_05_01.operations.ResourceGroupsOperations>`
+           * 2019-03-01: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2019_03_01.operations.ResourceGroupsOperations>`
            * 2019-05-01: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2019_05_01.operations.ResourceGroupsOperations>`
            * 2019-05-10: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2019_05_10.operations.ResourceGroupsOperations>`
            * 2019-07-01: :class:`ResourceGroupsOperations<azure.mgmt.resource.resources.v2019_07_01.operations.ResourceGroupsOperations>`
@@ -306,6 +323,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import ResourceGroupsOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import ResourceGroupsOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import ResourceGroupsOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import ResourceGroupsOperations as OperationClass
         elif api_version == '2019-05-10':
@@ -331,6 +350,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2017_05_10.operations.ResourcesOperations>`
            * 2018-02-01: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2018_02_01.operations.ResourcesOperations>`
            * 2018-05-01: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2018_05_01.operations.ResourcesOperations>`
+           * 2019-03-01: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2019_03_01.operations.ResourcesOperations>`
            * 2019-05-01: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2019_05_01.operations.ResourcesOperations>`
            * 2019-05-10: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2019_05_10.operations.ResourcesOperations>`
            * 2019-07-01: :class:`ResourcesOperations<azure.mgmt.resource.resources.v2019_07_01.operations.ResourcesOperations>`
@@ -349,6 +369,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import ResourcesOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import ResourcesOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import ResourcesOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import ResourcesOperations as OperationClass
         elif api_version == '2019-05-10':
@@ -374,6 +396,7 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
            * 2017-05-10: :class:`TagsOperations<azure.mgmt.resource.resources.v2017_05_10.operations.TagsOperations>`
            * 2018-02-01: :class:`TagsOperations<azure.mgmt.resource.resources.v2018_02_01.operations.TagsOperations>`
            * 2018-05-01: :class:`TagsOperations<azure.mgmt.resource.resources.v2018_05_01.operations.TagsOperations>`
+           * 2019-03-01: :class:`TagsOperations<azure.mgmt.resource.resources.v2019_03_01.operations.TagsOperations>`
            * 2019-05-01: :class:`TagsOperations<azure.mgmt.resource.resources.v2019_05_01.operations.TagsOperations>`
            * 2019-05-10: :class:`TagsOperations<azure.mgmt.resource.resources.v2019_05_10.operations.TagsOperations>`
            * 2019-07-01: :class:`TagsOperations<azure.mgmt.resource.resources.v2019_07_01.operations.TagsOperations>`
@@ -392,6 +415,8 @@ class ResourceManagementClient(MultiApiClientMixin, SDKClient):
             from .v2018_02_01.operations import TagsOperations as OperationClass
         elif api_version == '2018-05-01':
             from .v2018_05_01.operations import TagsOperations as OperationClass
+        elif api_version == '2019-03-01':
+            from .v2019_03_01.operations import TagsOperations as OperationClass
         elif api_version == '2019-05-01':
             from .v2019_05_01.operations import TagsOperations as OperationClass
         elif api_version == '2019-05-10':