+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/resources/v2020_06_01/operations/_operations.py
@@ -48,9 +48,9 @@ class Operations(object):
          deserialized response
         :param operation_config: :ref:`Operation configuration
          overrides<msrest:optionsforoperations>`.
-        :return: An iterator like instance of Operation
+        :return: An iterator like instance of Operation1
         :rtype:
-         ~azure.mgmt.resource.resources.v2020_06_01.models.OperationPaged[~azure.mgmt.resource.resources.v2020_06_01.models.Operation]
+         ~azure.mgmt.resource.resources.v2020_06_01.models.Operation1Paged[~azure.mgmt.resource.resources.v2020_06_01.models.Operation1]
         :raises: :class:`CloudError<msrestazure.azure_exceptions.CloudError>`
         """
         def prepare_request(next_link=None):
@@ -96,7 +96,7 @@ class Operations(object):
         header_dict = None
         if raw:
             header_dict = {}
-        deserialized = models.OperationPaged(internal_paging, self._deserialize.dependencies, header_dict)
+        deserialized = models.Operation1Paged(internal_paging, self._deserialize.dependencies, header_dict)
 
         return deserialized
     list.metadata = {'url': '/providers/Microsoft.Resources/operations'}