+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/subscriptions/v2016_06_01/models/_subscription_client_enums.py
@@ -26,3 +26,9 @@ class SpendingLimit(str, Enum):
     on = "On"
     off = "Off"
     current_period_off = "CurrentPeriodOff"
+
+
+class ResourceNameStatus(str, Enum):
+
+    allowed = "Allowed"
+    reserved = "Reserved"