+++ sdk/resources/azure-mgmt-resource/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 10.2.0 (2020-08-03)
+
+**Features**
+
+  - Model AliasPath has a new parameter metadata
+  - Model Alias has a new parameter default_metadata
+  - Model ProviderResourceType has a new parameter default_api_version
+  - Model ProviderResourceType has a new parameter api_profiles
+  - Added operation group SubscriptionClientOperationsMixin
+
 ## 10.1.0 (2020-06-30)
 
 **Features**